requirements.txt
flask==3.0.3
line-bot-sdk==3.12.0
pandas==2.2.2
openpyxl==3.1.3
gunicorn==22.0.0
